FontDraw関数、NumConversion関数完成。 ObjMessageを使いやすくした。('；'を入れなくてOK) ミサイルリキャストタイムのグローバル変数が何故かintになっていたのでfloatに修正。 フォント入れ替え作業等で一切使わない変数が宣言されたままだったのをコメントアウトしたり、 小数点の後に"f"つけ忘れ部分をつけたり等の微修正。

+++ 星育喰/星育喰/ObjTitle.cpp
@@ -54,7 +54,7 @@ int g_Speed_num = 1000;
 int g_Balance_num = 1000;
 int g_Research_num = 0;
 int g_Remain_num = 6000;
-int g_Recast_time = 3;
+float g_Recast_time = 3.0f;
 int g_P_Planet_damage = 0;
 
 int g_Iron_num = 0;
